requirements.txt
matplotlib==3.6.2
pandas==1.5.2
Pillow==10.0.1
Pillow==10.1.0
plotly==5.17.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.27.2
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
